dseva-backend/collector/requirements.txt
PyGithub==2.3.0
requests==2.32.3
debugpy==1.5.1